requirements.txt
hindsight-client
openai
python-dotenv
streamlit